+++ core/src/com/unciv/ui/saves/LoadGameScreen.kt
@@ -7,8 +7,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
 import com.badlogic.gdx.utils.SerializationException
 import com.unciv.Constants
-import com.unciv.logic.UncivFiles
 import com.unciv.logic.MissingModsException
+import com.unciv.logic.UncivFiles
 import com.unciv.logic.UncivShowableException
 import com.unciv.models.ruleset.RulesetCache
 import com.unciv.models.translations.tr
@@ -240,7 +240,7 @@ class LoadGameScreen(previousScreen:BaseScreen) : LoadOrSaveScreen() {
                     val repo = repos.items.firstOrNull { it.name.lowercase() == modName }
                         ?: throw UncivShowableException("Could not find a mod named \"[$modName]\".")
                     val modFolder = Github.downloadAndExtract(
-                        repo.html_url, repo.default_branch,
+                        repo,
                         Gdx.files.local("mods")
                     )
                         ?: throw Exception() // downloadAndExtract returns null for 404 errors and the like -> display something!