+++ core/src/com/unciv/ui/utils/extensions/Scene2dExtensions.kt
@@ -317,7 +317,7 @@ fun Actor.surroundWithCircle(size: Float, resizeActor: Boolean = true, color: Co
 fun Actor.addBorder(size:Float, color: Color, expandCell:Boolean = false): Table {
     val table = Table()
     table.pad(size)
-    table.background = ImageGetter.getBackground(color)
+    table.background = BaseScreen.skinStrings.getUiBackground("General/Border", tintColor = color)
     val cell = table.add(this)
     if (expandCell) cell.expand()
     cell.fill()